refactor(frontend): reorganize components into feature-based folder structure
