+++ frontend/app/layout.tsx
@@ -4,9 +4,11 @@ import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 
 import { Toaster } from '@/components/ui/sonner';
-import './globals.css';
 
-import { AuthProvider } from '@/components/authProvider';
+// @ts-expect-error - CSS import
+import '@/app/globals.css';
+
+import { AuthProvider } from '@/components/shared/authProvider';
 import { TooltipProvider } from '@/components/ui/tooltip';
 
 const geistSans = Geist({
