+++ frontend/components/friends/acceptFriendshipDialog.tsx
