+++ frontend/components/friends/rejectFriendshipDialog.tsx
