+++ frontend/components/messages/conversationArea.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from './ui/button';
+import { Button } from '../ui/button';
 
 import Image from 'next/image';
 import { Icon } from '@iconify/react';
