+++ frontend/components/sidebar/sidebarPopover.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from './ui/button';
+import { Button } from '../ui/button';
 import {
   Popover,
   PopoverContent,
